Issue #2840050: Move update methods to LocalStockUpdater.
